quiet some debugging

+++ libraries/entities/src/EntityScriptingInterface.cpp
@@ -56,7 +56,11 @@ EntityItemID EntityScriptingInterface::identifyEntity(EntityItemID entityID) {
         // found it!
         entityID.id = actualID.id;
         entityID.isKnownID = true;
-        qDebug() << "EntityScriptingInterface::identifyEntity() ...found it... isKnownID=" << entityID.isKnownID << "id=" << entityID.id << "creatorTokenID=" << entityID.creatorTokenID;
+        bool wantDebug = false;
+        if (wantDebug) {
+            qDebug() << "EntityScriptingInterface::identifyEntity() ...found it... isKnownID=" << entityID.isKnownID 
+                            << "id=" << entityID.id << "creatorTokenID=" << entityID.creatorTokenID;
+        }
     }
     return entityID;
 }
@@ -135,10 +139,13 @@ void EntityScriptingInterface::deleteEntity(EntityItemID entityID) {
         entityID.id = actualID.id;
         entityID.isKnownID = true;
 
-        qDebug() << "EntityScriptingInterface::deleteEntity()... " 
-                    << "isKnownID=" << entityID.isKnownID 
-                    << "id=" << entityID.id 
-                    << "creatorTokenID=" << entityID.creatorTokenID;
+        bool wantDebug = false;
+        if (wantDebug) {
+            qDebug() << "EntityScriptingInterface::deleteEntity()... " 
+                        << "isKnownID=" << entityID.isKnownID 
+                        << "id=" << entityID.id 
+                        << "creatorTokenID=" << entityID.creatorTokenID;
+        }
 
         getEntityPacketSender()->queueEraseEntityMessage(entityID);
     }
@@ -161,7 +168,11 @@ EntityItemID EntityScriptingInterface::findClosestEntity(const glm::vec3& center
         if (closestEntity) {
             result.id = closestEntity->getID();
             result.isKnownID = true;
-            //qDebug() << "EntityScriptingInterface::findClosestEntity()... isKnownID=" << result.isKnownID << "id=" << result.id << "creatorTokenID=" << result.creatorTokenID;
+            bool wantDebug = false;
+            if (wantDebug) {
+                qDebug() << "EntityScriptingInterface::findClosestEntity()... isKnownID=" << result.isKnownID 
+                            << "id=" << result.id << "creatorTokenID=" << result.creatorTokenID;
+            }
         }
     }
     return result;
